Add OpenSSL installation before running linter for Ubuntu platforms
This change was made to ensure that the super-linter validation job on Ubuntu systems can successfully install OpenSSL and run the linter without errors. The problem was that the linter required OpenSSL libraries to be installed, but they were not included in the default Ubuntu image. By adding the installation of OpenSSL as a separate step before running the linter, we can now build and test our code on Ubuntu environments more reliably.

Note that this change only affects the CI workflow for Ubuntu platforms (as denoted by "ubuntu-latest" in the task description). The change was made to improve the consistency of the CI pipeline, as it is important that our builds and tests pass on all supported platforms.

Closes: #12